Add reviewer boolean to star document
When a PR is ninja starred we will check if that user is part of
the review team *with his own token*, and save a true/false
flag in the star document.

Pro: we do not need to ensure we have the proper access token  when
checking team membership in order to create the status.

Con: if a user on the review teamstars from GH, they *must* be a user
in RN in order for their star to count (we need their token).